services/draw/requirements.txt
fdk
boto3
pillow
